Add candidate applications

+++ client/src/app/site/pages/meetings/pages/assignments/pages/assignment-detail/components/assignment-detail/assignment-detail.component.ts
@@ -343,8 +343,12 @@ export class AssignmentDetailComponent extends BaseMeetingComponent implements O
      * @param candidate A ViewAssignmentUser currently in the list of related users
      */
     public async removeCandidate(candidate: ViewAssignmentCandidate): Promise<void> {
-        await this.assignmentCandidateRepo.delete(candidate);
-        this.updateCandidatesArray();
+        const name = this.getCandidateName(candidate);
+        const title = this.translate.instant(`Are you sure you want to remove this candidate?`);
+        if (await this.promptService.open(title, name)) {
+            await this.assignmentCandidateRepo.delete(candidate);
+            this.updateCandidatesArray();
+        }
     }
 
     private updateCandidatesArray(): void {
@@ -364,9 +368,19 @@ export class AssignmentDetailComponent extends BaseMeetingComponent implements O
     public async removeSelf(): Promise<void> {
         const candidate = this.assignment.candidates.find(c => c.user_id === this.operator.operatorId);
         if (candidate) {
-            await this.removeCandidate(candidate);
+            const title = this.translate.instant(`Are you sure you want to remove yourself as candidate?`);
+            const name = this.getCandidateName(candidate);
+            if (await this.promptService.open(title, name)) {
+                await this.assignmentCandidateRepo.delete(candidate);
+                this.updateCandidatesArray();
+            }
         }
     }
+
+    public getCandidateName(candidate: ViewAssignmentCandidate | null): string {
+        return candidate?.user?.short_name || candidate?.user?.full_name || candidate?.getTitle() || ``;
+    }
+
 
     /**
      * Triggers an update of the sorting.
